add almost working audio
should probably make ring buffer into an actual ring buffer, but otherwise it's pretty close

+++ src/audio.h
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 
-using AudioBuffer = std::vector<uint8_t>;
+using AudioBuffer = std::vector<float>;
 
 class Audio
 {
